refactor searchresults
